requirements.txt
setuptools
flatbuffers
python-dotenv
pycryptodome
xxhash
requests
cloudscraper
unitypy
pillow
click
tqdm
pydub
./PyCriCodecs
./crcmanip
